调整UI

+++ Assets/HotfixScripts/Runtime/HotfixBusiness/Common/LubanConfig/Cfg/ItemConfig.cs
@@ -19,10 +19,10 @@ public sealed partial class ItemConfig :  Bright.Config.BeanBase
         Type = _buf.ReadInt();
         Name = _buf.ReadString();
         Icon = _buf.ReadString();
-        Hp = _buf.ReadInt();
-        Spd = _buf.ReadInt();
-        Atk = _buf.ReadInt();
-        Def = _buf.ReadInt();
+        Hp = _buf.ReadFloat();
+        Spd = _buf.ReadFloat();
+        Atk = _buf.ReadFloat();
+        Def = _buf.ReadFloat();
         PostInit();
     }
 
@@ -50,19 +50,19 @@ public sealed partial class ItemConfig :  Bright.Config.BeanBase
     /// <summary>
     /// 血量修正
     /// </summary>
-    public int Hp { get; private set; }
+    public float Hp { get; private set; }
     /// <summary>
     /// 速度修正
     /// </summary>
-    public int Spd { get; private set; }
+    public float Spd { get; private set; }
     /// <summary>
     /// 攻击修正
     /// </summary>
-    public int Atk { get; private set; }
+    public float Atk { get; private set; }
     /// <summary>
     /// 防御修正
     /// </summary>
-    public int Def { get; private set; }
+    public float Def { get; private set; }
 
     public const int __ID__ = -783814593;
     public override int GetTypeId() => __ID__;
